requirements.txt
python-chess
numpy
tensorflow
scikit-learn